feat(Image): 타입스크립트 변환
1. 타입스크립트 적용

closes: #795
