+++ src/index.js
@@ -13,7 +13,7 @@ import Timeline from '@Components/charts/Timeline'
 import LineMergeTimeline from '@Components/charts/LineMergeTimeline'
 import Navbar from '@Components/layout/Navbar'
 import Footer from './components/layout/Footer'
-import Image from '@Components/layout/Image'
+import Image from './components/layout/Image'
 import Heading from './components/layout/Heading'
 import Tabs from '@Components/layout/Tabs'
 import Histogram from '@Components/charts/Histogram'
